% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{baseTheoffice}
\alias{baseTheoffice}
\title{baseTheoffice}
\format{
um data frame com 163 linhas e 9 variaveis:

\describe{
\item{temporada}{temporada do episodio.}
\item{n_total}{numero do episodio correspondente ao total de todas as temporadas.}
\item{n_episodio}{numero do episodio correspondente a temporada.}
\item{titulo}{nome do episodio.}
\item{dirigido_por}{diretor do episodio.}
\item{escrito_por}{escritor do roteiro do episodio.}
\item{audiencia}{numero da audiencia na data de exibicao original.}
\item{codigo}{codigo referente ao episodio.}
\item{exibicao_original}{a data da exibicao original do episodio.}
}
}
\source{
baseTheoffice
}
\usage{
baseTheoffice
}
\description{
Um conjunto de dados que contem informacoes sobre os episódios da serie The Office´.
}
\keyword{datasets}
